src/test/java/ru/netology/stats/StatsServiceTest.java
package ru.netology.stats;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

public class StatsServiceTest {
    @Test
    public void shouldSum() {
        StatsService service = new StatsService();

        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long fact = service.sum(sales);
        long expected = 8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18;

        Assertions.assertEquals(expected, fact);
    }

    @Test
    public void shouldFindMinMonth() {
        StatsService service = new StatsService();

        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long fact = service.minSales(sales);
        long expected = 9;

        Assertions.assertEquals(expected, fact);
    }

    @Test
    public void shouldFindMaxMonth() {
        StatsService service = new StatsService();

        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long fact = service.maxSales(sales);
        long expected = 8;

        Assertions.assertEquals(expected, fact);
    }

    @Test
    public void shouldAverage() {
        StatsService service = new StatsService();

        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        double fact = service.average(sales);
        double expected = 15;

        Assertions.assertEquals(expected, fact);
    }

    @Test
    public void shouldCountAboveAvg() {
        StatsService service = new StatsService();

        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long fact = service.aboveAvg(sales);
        int expected = 5;

        Assertions.assertEquals(expected, fact);
    }

    @Test
    public void shouldCountBelowAvg() {
        StatsService service = new StatsService();

        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long fact = service.belowAvg(sales);
        int expected = 5;

        Assertions.assertEquals(expected, fact);
    }

}
